Color deprecation warnings
Also:

- Move all the warning related parts into a separate file.
