+++ unittests/test_loader.py
@@ -8,9 +8,9 @@ import pytest
 
 import reframe as rfm
 from reframe.core.exceptions import (ConfigError, NameConflictError,
-                                     ReframeDeprecationWarning,
                                      RegressionTestLoadError)
 from reframe.core.systems import System
+from reframe.core.warnings import ReframeDeprecationWarning
 from reframe.frontend.loader import RegressionCheckLoader
